requirements.txt
# Add your project dependencies here
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml
google-api-python-client
